study cafe api 추가

+++ src/main/java/com/seatly/seatly/dto/studycafe/StudyCafeUsage.java
@@ -7,7 +7,7 @@ import lombok.Data;
 @AllArgsConstructor
 public class StudyCafeUsage {
 
-  private Long totalCount;
-  private Long useCount;
+  private int totalCount;
+  private int useCount;
 
 }
